@@ -1,19 +1,45 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netstandard2.0</TargetFramework>
-    <Version>1.4.0</Version>
-    <AssemblyVersion>1.4.0.0</AssemblyVersion>
-    <LangVersion>7.1</LangVersion>
-    <FileVersion>1.4.0.0</FileVersion>
+    <TargetFramework>netstandard2.1</TargetFramework>
+    <LangVersion>8.0</LangVersion>
+    <Nullable>enable</Nullable>
+    <WarningsAsErrors>NU1605;CS8625;CS8600;CS8602</WarningsAsErrors>
+    <AssemblyOriginatorKeyFile>..\..\..\syp.biz.SockJS.NET.pfx</AssemblyOriginatorKeyFile>
+    <SignAssembly>true</SignAssembly>
+    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
+    <AssemblyVersion>2.0.0.0</AssemblyVersion>
+    <FileVersion>2.0.0.0</FileVersion>
+  </PropertyGroup>
+
+  <PropertyGroup>
+    <PackageId>syp.biz.SockJS.NET.Client</PackageId>
+    <Description>An asynchronous .NET implementation of the SockJS client</Description>
+    <Product>SockJS.NET.Client</Product>
+    <Version>2.0.0</Version>
+    <Authors>syp.biz</Authors>
+    <Company>syp.biz</Company>
+    <PackageProjectUrl>https://github.com/sypbiz/SockJS.NET</PackageProjectUrl>
+    <RepositoryUrl>https://github.com/sypbiz/SockJS.NET</RepositoryUrl>
+    <PackageRequireLicenseAcceptance>true</PackageRequireLicenseAcceptance>
+    <PackageLicenseFile>LICENSE</PackageLicenseFile>
+    <Copyright>Copyright (c) 2020 syp.biz</Copyright>
+    <PackageTags>SockJS; Client; .NET</PackageTags>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Newtonsoft.Json" Version="11.0.2" />
+    <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
   </ItemGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\syp.biz.SockJS.NET.Common\syp.biz.SockJS.NET.Common.csproj" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <None Include="..\..\..\LICENSE">
+      <Pack>True</Pack>
+      <PackagePath></PackagePath>
+    </None>
   </ItemGroup>
 
 </Project>
