@@ -1,9 +1,17 @@
-﻿namespace syp.biz.SockJS.NET.Common.Interfaces
+﻿using System;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace syp.biz.SockJS.NET.Common.Interfaces
 {
-    public interface ITransport : IEventEmitter
+    public interface ITransport : IDisposable
     {
-        string TransportName { get; }
-        void Close();
-        void Send(string message);
+        event EventHandler<string> Message;
+        event EventHandler Disconnected;
+        
+        Task Connect(CancellationToken token);
+        Task Disconnect();
+
+        Task Send(string data, CancellationToken token);
     }
 }
