@@ -1,21 +1,33 @@
 ﻿using System;
+using System.Diagnostics.CodeAnalysis;
 using System.Linq;
+using System.Text;
 
 namespace syp.biz.SockJS.NET.Common.Utils
 {
+    [SuppressMessage("ReSharper", "UnusedType.Global")]
+    [SuppressMessage("ReSharper", "UnusedMember.Global")]
     public static class Random
     {
-        private const string RANDOM_STRING_CHARS = "abcdefghijklmnopqrstuvwxyz012345";
+        private const string RandomStringChars = "abcdefghijklmnopqrstuvwxyz012345";
 
         public static string GetString(int length)
         {
             if (length < 1) throw new ArgumentOutOfRangeException(nameof(length), "Length must be at least 1");
 
             var buffer = new byte[length];
+            var charsLength = RandomStringChars.Length;
+
             System.Security.Cryptography.RandomNumberGenerator.Create().GetBytes(buffer);
-            return string.Join(string.Empty, buffer.Select(b => RANDOM_STRING_CHARS.Substring(b % length, 1)));
+
+            var chars = buffer
+                .Select(b => b % charsLength) // relative position (index) of random byte in RandomStringChars
+                .Select(idx => RandomStringChars[idx]); // select char for index
+
+            return new StringBuilder(length).Append(chars).ToString();
         }
 
+        [SuppressMessage("ReSharper", "MemberCanBePrivate.Global")]
         public static long GetNumber(long max) => (long) Math.Floor(new System.Random(DateTime.UtcNow.Millisecond).NextDouble() * max);
 
         public static string GetNumberString(long max)
